pat2prism/requirements.txt
jinja2>=3.0
antlr4-python3-runtime>=4.9.2
pytest>=7.0
